+++ ci/cpu/build.sh
@@ -5,20 +5,19 @@
 ######################################
 set -e
 
-# Logger function for build status output
-function logger() {
-  echo -e "\n>>>> $@\n"
-}
-
 # Set path and build parallel level
-export PATH=/conda/bin:/usr/local/cuda/bin:$PATH
-export PARALLEL_LEVEL=4
+export PATH=/opt/conda/bin:/usr/local/cuda/bin:$PATH
+export PARALLEL_LEVEL=${PARALLEL_LEVEL:-4}
 
 # Set home to the job's workspace
 export HOME=$WORKSPACE
 
 # Switch to project root; also root of repo checkout
 cd $WORKSPACE
+
+# Setup 'gpuci_conda_retry' for build retries (results in 2 total attempts)
+export GPUCI_CONDA_RETRY_MAX=1
+export GPUCI_CONDA_RETRY_SLEEP=30
 
 # If nightly build, append current YYMMDD to version
 if [[ "$BUILD_MODE" = "branch" && "$SOURCE_BRANCH" = branch-* ]] ; then
@@ -29,17 +28,22 @@ fi
 # SETUP - Check environment
 ################################################################################
 
-logger "Get env..."
+gpuci_logger "Get env"
 env
 
-logger "Activate conda env..."
-source activate gdf
+gpuci_logger "Activate conda env"
+. /opt/conda/etc/profile.d/conda.sh
+conda activate rapids
 
-logger "Check versions..."
+gpuci_logger "Check versions"
 python --version
 gcc --version
 g++ --version
-conda list
+
+gpuci_logger "Check conda environment"
+conda info
+conda config --show-sources
+conda list --show-channel-urls
 
 # FIX Added to deal with Anancoda SSL verification issues during conda builds
 conda config --set ssl_verify False
@@ -47,26 +51,30 @@ conda config --set ssl_verify False
 ################################################################################
 # BUILD - Conda package builds (conda deps: librmm <- rmm)
 ################################################################################
+
 if [[ "$BUILD_LIBRMM" == "1" ]]; then
-  logger "Build conda pkg for librmm..."
-  source ci/cpu/librmm/build_librmm.sh
+  gpuci_logger "Build conda pkg for librmm"
+  if [[ -z "$PROJECT_FLASH" || "$PROJECT_FLASH" == "0" ]]; then
+    conda build conda/recipes/librmm --python=$PYTHON
+  else
+    conda build --dirty --no-remove-work-dir conda/recipes/librmm
+  fi
 fi
 
 if [[ "$BUILD_RMM" == "1" ]]; then
-  logger "Build conda pkg for rmm..."
-  source ci/cpu/rmm/build_rmm.sh
+  gpuci_logger "Build conda pkg for rmm"
+  if [[ -z "$PROJECT_FLASH" || "$PROJECT_FLASH" == "0" ]]; then
+    conda build conda/recipes/rmm --python=$PYTHON
+  else
+    conda build --dirty --no-remove-work-dir \
+      -c $WORKSPACE/ci/artifacts/rmm/cpu/conda-bld/ conda/recipes/rmm
+  fi
 fi
 
 ################################################################################
 # UPLOAD - Conda packages
 ################################################################################
 
-if [[ "$BUILD_LIBRMM" == "1" ]]; then
-  logger "Upload conda pkg for librmm..."
-  source ci/cpu/librmm/upload-anaconda.sh
-fi
+gpuci_logger "Upload conda packages"
+source ci/cpu/upload.sh
 
-if [[ "$BUILD_RMM" == "1" ]]; then
-  logger "Upload conda pkg for rmm..."
-  source ci/cpu/rmm/upload-anaconda.sh
-fi