+++ CHANGELOG.md
@@ -12,12 +12,14 @@
 - PR #612 Drop old Python `device_array*` API
 - PR #603 Always test both legacy and per-thread default stream
 - PR #611 Add a note to the contribution guide about requiring 2 C++ reviewers
+- PR #615 Improve gpuCI Scripts
 
 ## Bug Fixes
 
 - PR #592 Add `auto_flush` to `make_logging_adaptor`
 - PR #602 Fix `device_scalar` and its tests so that they use the correct CUDA stream
 - PR #621 Make `rmm::cuda_stream_default` a `constexpr`
+- PR #625 Use `librmm` conda artifact when building `rmm` conda package
 
 # RMM 0.16.0 (21 Oct 2020)